refactor: cognito client in new file
